Applied patch from Stefan Kost
when running lcov over an uninstalled user-space apps tests, it finds
the .da file in the .libs directories, but does not look for the sources
one hierarchy up. Libtool places the object in the .libs dirs. when
running gcov manually one can specify -o.libs/ to produce a source.c.gov
file. I now have attached a patch that fixes the problem for me. please
do not just ignore this report. the lcov tool is so nice and it would be
a shame if it can not be used for normal apps.