lcov: adding support for gzipped html
... based on patch by dnozay@vmware.com

dnozay@vmware.com: genhtml is a great tool to generate html, but the more
files, the more space it takes (here I have over 113MB of html generated),
add to that I need to have different sets, and space usage increases
dramatically (2.7GB). we are using browsers with htmlz support, so it would
be nice to have support for that in genhtml, relying on 'gzip -S z' to do
the job.