lcov: geninfo chokes on spaces in the directory name
In lcov 1.6, geninfo fails to find gcno/gcda files if the source directory
has spaces in the name, because it uses backticks to shell out to "find",
passing $directory on the command-line.

Attached is a patch that double-quotes the variable, allowing geninfo to
operate on directories with spaces in their name. The fix isn't perfect; it
won't work on directories with a " character, but it works fine for my
purposes (I don't have any directories with quotes). A better fix would be
to use IPC::System::Simple's capturex from
http://search.cpan.org/~pjf/IPC-System-Simple-0.15/lib/IPC/System/Simple.pm
#runx(),_systemx()_and_capturex(). capturex() is a multiple-argument form
of the backticks, so it avoids any interpolation errors.