lcov: Sort branches in unnamed blocks first
When processing branch coverage data, consider branches in "unnamed"
blocks to come before other blocks on the same line, so that they
appear in the correct order in HTML output.

This is accomplished by using block number -1 for unnamed blocks,
instead of 9999 as was previously done.  In branch data vectors, this
is reprsented by the value $BR_VEC_MAX, which is defined to be the
largest value representable in the field width used.  This same value
is also used in .info files, for backward-compatibility with regular
expressions used to parse these files.  As a result, .info files
generated by versions of lcov with this change can be read by older
versions, though branch results will still appear out of order.

Signed-off-by: Jeffrey Hutzelman <jhutz@cmu.edu>