lcov: Fix output files being created in / directory
When a warning is emitted by lcov before creating an output file,
e.g. when a negative coverage count was found while combining
tracefiles, lcov tries to create the output file in the root
directory (/) instead of the current working directory.

This is a result of lcov's warn handler calling a temp file cleanup
routine that changes directories to / before trying to remove its
temporary directory.

Fix this by removing the temp cleanup call from the warn handler.

Signed-off-by: Peter Oberparleiter <oberpar@linux.vnet.ibm.com>