genhtml: Fix warning with small genhtml_line_field_width
On systems with Perl versions 5.21 and above, genhtml prints a warning
similar to the following during processing:

  genhtml: Negative repeat count does nothing at bin/genhtml line 3854,
           <SOURCE_HANDLE> line 4.

This is due to size calculations resulting in a negative number of
padding characters when genhtml_line_field_width is lower than the size
of the strings to pad (9). Fix this by disabling padding in these cases.

Reported-by: xaizek@openmailbox.org
Signed-off-by: Peter Oberparleiter <oberpar@linux.vnet.ibm.com>