Add options to choose demangler and --no-strip-underscores
Using lcov in cygwin to generate coverage for Mingw-based compilers, I needed to change the demangler (in order to use the c++filt.exe provided by mingw instead of the cygwin one) and explicitly request for "--no-strip-underscores" under a OS which is not darwin.

This patch makes both change available as genhtml options. I had no previous experience with perl, so it might contain some errors.