genhtml: honor the SOURCE_DATE_EPOCH variable
Implement the SOURCE_DATE_EPOCH specification[1] for reproducible
builds. If SOURCE_DATE_EPOCH is set, use it as timestamp instead of the
current time.

In this context, reproducible builds means reproducible HTML coverage
reports.

Known bug: the specification[1] says to defer converting the timestamp
to local timezone at presentation time. This is currently not happening;
it's converted at build time.

[1] https://reproducible-builds.org/specs/source-date-epoch/

Signed-off-by: Bjørn Forsman <bjorn.forsman@gmail.com>