geninfo: Fix data collection for files containing templates
When using gcc 8, lcov/geninfo produces corrupt coverage output for
source code that contains templates or other constructs that cause gcov
to produce multiple versions of output for some lines and branches.

This is caused by an incorrect check for duplicate output in function
read_gcov_file() that is triggered when a template consists of multiple
lines, or contains branches.

Fix this by ensuring that duplicate lines in per-instance gcov output are
correctly ignored. Only the initial occurrence of each line containing
the combined coverage of all instances will be processed by geninfo.

Note that for branch coverage, gcov doesn't provide a combined view and
geninfo processes all branches provided. This should not be a problem
though as genhtml will combine the branch data when generating HTML
output.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>