geninfo: preserve-paths makes gcov to fail for long pathnames
geninfo uses '--preserve-paths' gcov option whenever gcov supports it, this
forces gcov to use a whole pathname as a filename for .gcov files. So in cases
of quite large pathnames, gcov isn't able to create .gcov files and hence
geninfo can't get any data. The fix replaces usage '--preserve-paths' with
'--hash-filenames' when it is available.

Signed-off-by: Igor Ignatev <igor.v.ignatiev@gmail.com>