geninfo: Fix errors while resolving /././ path components
Trying to collect code coverage data for source code that contains
repeat ./ references in a path components fails with the following
error message:

  geninfo: WARNING: cannot find an entry for <name>.gcov in .gcno file,
           skipping file!

This is caused by a bug in path normalization function
solve_relative_path() that does not correctly process adjacent ./
references.

Fix this by repeating the resolution of ./ references in path
components.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>
Reported-by: Joshua Cranmer