geninfo: Fix missing FN: entries in result files
geninfo sometimes fails to correctly collect function starting lines for
some source files, resulting in output files with missing FN: lines.
Also such functions are missing from the function list in HTML output.

The problem occurs when
  a) multiple source files contribute to a function implementation (e.g.
     via including code), and
  b) the source file that contains the initial function definition
     is not the source file that contains the most function
     definitions

The problem occurs due to a heuristic in function graph_find_base() that
incorrectly determines the source file for a function in this situation.

Fix this by using the first file that contributes to a function as the
base source file for that function. Only apply this change to data
collected using GCC versions 4 and above since earlier versions did not
produce stable file orders in graph files.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>
Reported-by: Joshua Cranmer