genhtml: Maintain order of branches in unnamed blocks
Branch data that is reported by gcov before a block number is shown in
genhtml output before all other branches for that line. This can lead to
confusion when trying to correlated branch coverage data with code
constructs.

This problem occurs because lcov assigns a '-1' as block number for
this branch data. During internal processing this value is converted to
2^32-1. When genhtml creates branch coverage output it sorts these
branches last.

Fix this by converting block numbers 2^32-1 back to -1 when reading
coverage data in genhtml

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>
Reported-by: mueller_ja@gmx.net