bin: remove outdated FSF mailing address from license boilerplate
The FSF address used in the license boilerplate text is no longer
correct since their offices have relocated. Rather than introduce
the new mailing address which will inevitably change again in the
future, use the website address which is more permanent.

The main COPYING file already had the correct mailing address and
is not desirable to change that to the URL since it is a legal
document.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>