Add configuration options for LCOV markers
The `LCOV_*_START` and `LCOV_*_STOP` markers were hard-coded in
`bin/geninfo`.  It was not possible to override them, e.g. in `lcovrc`.
This commit does not alter the default character sequences for these
markers, but extends reading the configuration to allow for overriding a
marker's character sequence.  The man page for `lcovrc` is updated
accordingly.