Enable exclusion of coverpoints within function body (line, branch, and the function itself) - controlled by a new "--erase-function regexp" option. Option is called "--erase-function" so as to not conflict with "--exclude", which had already been claimed for file exclusions via glob patterns.
In differential categorization, check for source changes in function
region:  if code has been added or removed, consider the function to
have been changed.  For changed functions, uncovered in current is
categorized as "UNC" and covered in current is categorized as "GNC".
If covered in baseline and not covered in current, keep categorization
as "LBC" (i.e., don't change to "UNC").

Add 'genhtml --show-proportion' option - default off.
If enabled, show proportion of line and branch coverpoints which are
exercised in "function coverage detail" table.
Detail tabled can be sorted by number of un-exercised coverpoints (as
well as alphabetically by function name, or by number of times function
is hit).  The sorted views may be useful - e.g., to help developer focus
attention on functions which would have the largest effect.

Add "genhtml --suppress-aliases" option - default off.
If enabled, suppress aliases in the 'function detail table' - show only
the leader.  This enables user to focus on the overall coverage which
can otherwise be lost in a sea of alias names, expecially for widely
instantiated templates.

These features resuire lcov/geninfo/genhtml to keep track of function
end lines - which requires gcov 9 or newer to do so.  Warn if the user's
gcov is too old.

Signed-off-by:  Henry Cox <henry.cox@mediatek.com>