lcov: Unclutter error and warn messages
Perl adds an 'at <sourcefile> line <linenumber>' suffix to all die/warn
messages that are not newline-terminated, making these messages harder
to read. As this type of information is irrelevant to most users and only
useful in the context of debugging problems within lcov, hide this
suffix unless --debug was specified.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>