geninfo: Fix external source file handling
When geninfo is called with '--no-external --directory ./' for source
files that are compiled with a toolchain that reports relative source
code path in gcov files, all source files will erroneously be marked as
external, resulting in output as follows:

  Excluded data for 1 file due to include/exclude options
  geninfo: WARNING:  no data generated

Affected are all toolchains that report relative source file paths in
gcov output (such as LLVM up to at least v16).

This error is due to geninfo external path matching sometimes looking
for 2 consecutive slashes in source code paths when the list of source
code base paths contain './' or '/'. Fix this by adjusting the matching
to require all base paths to contain a trailing slash. Also fix relative
path resolution failing for the '/' path.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>