genhtml: Fix anchor IDs
HTML 4.01 transitional requires [1] that anchor IDs start with a letter,
and may only contain characters of a restricted set. genhtml uses
numbers as IDs for line and date bin anchors, and might use IDs with
unsupported characters for user name and test name anchors.

Fix this by prefixing anchor IDs with a letter depending in the type of
anchor, and by replacing unsupported characters in user and test names.

The following anchro prefixes are introduced:

  - L for line anchors
  - B for date bin anchors
  - N for owner bin anchors
  - T for test name anchors

[1] https://www.w3.org/TR/html401/types.html#type-name

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>