genhtml: Fix file view TLA link style
genhtml generates nested <a id><a href> sequences that are not supported
by HTML 4.01 transitional. As a result the background style for the
nested link in TLA-classified branches is not always visible.

Fix this by using an enclosing <span> tag for defining line anchors.
Also use a defined foreground color for TLA-classified links to ensure a
working level of contrast between text and background.

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>