tests: use $CC and $CXX, not hard-coded gcc/g++
Cross-compiling may need to use a different compiler, so don't hard code
references to plain `gcc` and instead determine the compiler to use from
the environment.

When building for ChromeOS for example, invoking plain gcc or g++ is
turned into a fatal error as a safety measure to ensure that everything
which should be cross-compiled actually is; ignoring CC and CXX in the
environment causes the tests to break `make install`, since the test
scripts run unprefixed `gcc` to probe the compiler version in use even
when run as part of `make clean` invocations during `make install`.

Signed-off-by: Peter Marheine <pmarheine@chromium.org>