Check for consistency between baseline coverage DB, current coverage DB, and 'diff' file data: If file is present in both baseline and current coverage data:   - if version is not the same:  expect to find file in 'diff' data.   - if version is the same:  expect to NOT find in 'diff' data.
If file is not in 'current' DB: then it is no longer part of the project -
so we dont' care about differences.
If file is not present in 'baseline':  then it was recently added to the
project - so we are not looking at differences.

Checks motivated by report in issue #315.

Signed-off-by: Henry Cox <henry.cox@mediatek.com>