fix aggressive "rm -rf" with a less aggressive "rm -rf"
`make uninstall` deletes `$(PREFIX)/share` instead of
`$(PREFIX)/share/lcov` with `rm -rf`. This is surprising. This patch
fixes that unconditional delete by deleting `$(PREFIX)/share/lcov`. This
is still not optimal, and wouldn't be necessary if the tests were not
installed.

This patch also cleans up the installed manpages.

Signed-off-by: nega <nega0@users.noreply.github.com>