'genhtml --simplfy-script callback ..' feature. Use callback to munge function names for display. Goal is to compress the function detail table by, for example, shortening long template or namespace names.
Implemented as a callback rather than, for example, as a list of regexps because any project complicated enough to need simplification is complicated enough that the required regexps will be very, very complicated.
A callback is easier and more flexible.
See the sample implementation in .../scripts/simplify.pl

This feature may go part of the way toward satisfying #164.  (Function names will not be collapsed into one entry - but the table entries can be arbitrarily simple.)

Signed-off-by: Henry Cox <henry.cox@mediatek.com>